create new package for log tests, add new suite, update the test
